+++ benchmarkable/src/main/java/com/ocjadan/benchmarkable/questionDetails/QuestionDetails.kt
@@ -3,9 +3,9 @@ package com.ocjadan.benchmarkable.questionDetails
 import java.io.Serializable
 
 data class QuestionDetails(
-    val id: Int,
+    val id: Long,
     val isAnswered: Boolean,
-    val ownerId: Int,
+    val ownerId: Long,
     val ownerName: String,
     val ownerProfileImage: String
 ) : Serializable
